# core/txbuilder: improve signature witness signing perf