Working on improving TransferWithBlocks, and this is a tiny fix. 

Before
```
BenchmarkTransferWithBlocks-4   	     200	  31896144 ns/op	   61624 B/op	     919 allocs/op
```

After
```
BenchmarkTransferWithBlocks-4   	     300	  25861102 ns/op	   67099 B/op	     913 allocs/op
```