@kr 2016-12-16
For easy comparison, you can use [benchcmp](https://godoc.org/golang.org/x/tools/cmd/benchcmp) to produce output like this:

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkTransferWithBlocks-4     31896144      25861102      -18.92%

benchmark                         old bytes     new bytes     delta
BenchmarkTransferWithBlocks-4     61624         67099         +8.88%
```