@kr 2017-01-09
LGTM (sorry I neglected to comment before)

nit: imperative subject